requirements.txt
opencv-python
pytesseract
ultralytics
onvif-zeep
configparser
flask
